MAINT: Reduce PRIMA binary size and suppress warning
Incorporate changes to PRIMA that allow us to not compile the other
solvers. This reduces the binary from about 1.2MB to 840kB.

Also the warning about compact unwind is suppressed. It comes from using
AppleClang to link code compiled with gfortran. In theory it impacts
exception handling but tests with exception in the objective function,
constraint function, and in the C++ have shown no adverse effects.